Finished temperature regulation
Build the device, works well. Currently curved a bit conservatively.
Output PWM = 100% overdrives the fan because the input voltage is
greater than 12VDC. However, I think that could be okay. I would rather
have the curves a bit conservative in the beginning and if they are
annoyingly careful (i.e., too loud) I might tinker with it again
(currently the case is still open for me to fiddle with it).